+++ CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
 
 - Update to Swift 6.3.
 - Bump up supported macOS version to 15.
+- Deprecate `getHue(hue:saturation:lightness:alpha:)` in favor of `hslComponents`.
 
 
 ### fix